requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
atomicwrites==1.2.1       # via pytest
attrs==18.2.0             # via pytest
click==7.0
coverage==4.5.2           # via pytest-cov
django==1.11.16
mock==2.0.0
more-itertools==4.3.0     # via pytest
pbr==5.1.1
pluggy==0.8.0             # via pytest
py==1.7.0                 # via pytest
pytest-cov==2.6.0
pytest==4.0.1             # via pytest-cov
pytz==2018.7
pyyaml==3.13
six==1.11.0
stevedore==1.30.0
